Prune warning from clang about unused debug option.
LLVM emits a warning if the flag for debugging information '-g' or any
of its variants (-g1, -g2, -g3, etc) will not be used during the
particular invocation of LLVM, so for example, if LLVM is being used
simply to perform a link.

This commits adds this particular warning to the prune_warnings
function.

ChangeLog:

	* lib/target.exp (prune_warnings): Remove clang warning about
	unused debug option.